+++ tools/hie.nix
@@ -3,7 +3,7 @@
 pkgs.writeShellScript "hie.sh" ''
   set -e
   if [[ $1 == "--nh-help" ]]; then
-    echo "usage: nix run github:tbidne/nix-hs-tools#hie"
+    echo "usage: hie"
     exit 0
   fi
   ${pkgs.haskellPackages.implicit-hie}/bin/gen-hie > hie.yaml