+++ tools/ormolu.nix
@@ -8,7 +8,7 @@ pkgs.writeShellScript "ormolu.sh" ''
   dir=.
   while [ $# -gt 0 ]; do
     if [[ $1 == "--nh-help" ]]; then
-      echo "usage: nix run github:tbidne/nix-hs-tools#ormolu -- [--dir PATH] <args>"
+      echo "usage: ormolu [--dir PATH] <args>"
       exit 0
     elif [[ $1 == "--dir" ]]; then
       dir=$2